+++ spec/haml_lint/linter/object_reference_attributes_spec.rb
@@ -1,0 +1,23 @@
+require 'spec_helper'
+
+describe HamlLint::Linter::ObjectReferenceAttributes do
+  include_context 'linter'
+
+  context 'when a tag has no attributes' do
+    let(:haml) { '%tag' }
+
+    it { should_not report_lint }
+  end
+
+  context 'when a tag contains hash attributes' do
+    let(:haml) { "%tag{ lang: 'en' }" }
+
+    it { should_not report_lint }
+  end
+
+  context 'when a tag contains an object reference' do
+    let(:haml) { '%tag[@user]' }
+
+    it { should report_lint }
+  end
+end